refactors stars and brings it to App level
