+++ client/src/widgets/overview/ProductInfoTop.jsx
@@ -5,7 +5,7 @@
 import React, { useContext } from 'react';
 import { productContext } from '../../contexts/ProductContext.js';
 import { styleContext } from '../../contexts/StyleContext.js';
-import Stars from './Stars.jsx';
+import Stars from '../../Stars.jsx';
 
 function ProductInfoTop(props) {
   const { product } = useContext(productContext);
